fix authorization flaw

+++ controllers/stuff.js
@@ -62,13 +62,23 @@ module.exports.updateThing = (req, res, next) => {
     });
 };
 
+module.exports.deleteThing = (req, res, next) => {
+  Thing.findOne({ _id: req.params.id }).then((thing) => {
+    if (!thing) {
+      return res.status(404).json({ error: new Error("Not found ") });
+    }
+    if (thing.userId !== req.auth.userId) {
+      return res
+        .status(401)
+        .json({ error: new Error("Unauthorized activity") });
+    }
 
-module.exports.deleteThing = (req,res,next)=>{
-    Thing.deleteOne({_id:req.params.id})
-    .then(()=>{
-        res.status(200).json({message:"Deleted"})
-    })
-    .catch((error)=>{
-        res.status(400).json({error})
-    })
-}
+    Thing.deleteOne({ _id: req.params.id })
+      .then(() => {
+        res.status(200).json({ message: "Deleted" });
+      })
+      .catch((error) => {
+        res.status(400).json({ error });
+      });
+  });
+};
